@dscpinheiro 2022-02-17
@perf-cat This is just a personal repo I use for testing C#, .NET Core, and GitHub features. 

I'm not really looking for any external contributions at the moment.